requirements.txt
pillow==10.3.0
pyinstaller==6.7.0
tkinterdnd2==0.3.0
